mana points and souls

+++ Assets/Scripts/Managers/EventManager.cs
@@ -14,7 +14,7 @@ public class EventManager : MonoBehaviour
     public delegate void PlayerGetHitAction();
     public static PlayerGetHitAction OnPlayerGetHit;
 
-    public delegate void PlayerUseManaAction(float manaCost);
+    public delegate void PlayerUseManaAction(int manaCost);
     public static PlayerUseManaAction OnPlayerUseMana;
 
     public delegate void EnemyDeathAction();
